add popover on the map
