+++ app/assets/styles/core/_map.scss
@@ -71,4 +71,6 @@ p.map-layer__title {
   width: 100%;
   font-size: 0.875rem;
   margin-bottom: $global-spacing/4;
+
 }
+
